Rename command to 'migrate-api-version'
