+++ cmd/ctl/pkg/upgrade/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/upgrade",
     visibility = ["//visibility:public"],
     deps = [
-        "//cmd/ctl/pkg/upgrade/migrate:go_default_library",
+        "//cmd/ctl/pkg/upgrade/migrateapiversion:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
         "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
     ],
@@ -23,7 +23,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//cmd/ctl/pkg/upgrade/migrate:all-srcs",
+        "//cmd/ctl/pkg/upgrade/migrateapiversion:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
