+++ cmd/ctl/pkg/upgrade/migrateapiversion/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "command.go",
         "migrator.go",
     ],
-    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/upgrade/migrate",
+    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/upgrade/migrateapiversion",
     visibility = ["//visibility:public"],
     deps = [
         "//cmd/ctl/pkg/build:go_default_library",
